feat: add `derivedFrom` slices for Profiles
This adds slices on `derivedFrom` for all Profiles for Observations
which are derived from other Observations (eg Zscore, % pred, FEV1/FVC).

At the moment, this causes the IG build to raise the following error:
"Slicing cannot be evaluated: Profile based discriminators must have a
type with a profile (Reference.reference in profile
https://automatemedical.com/fhir/pulmonary-function-test/StructureDefinition/NAME"
where NAME is the name of the StructureDefinition (ie the Profile name).

This error should resolve when we have that canonical URL working.